fix comments from ivan
